third_party/WebKit/LayoutTests/fast/dom/idl-union-type-unittest.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description('IDL union types unittests');

if (window.internals) {
    // The conversion rule prefers string over double, hence a value which
    // is neither double nor string will be converted using ToString().

    var unionTypesTest = internals.unionTypesTest();

    debug('Tests for attributes');
    shouldBe('unionTypesTest.doubleOrStringAttribute', '0');
    unionTypesTest.doubleOrStringAttribute = 3.14;
    shouldBe('unionTypesTest.doubleOrStringAttribute', '3.14');
    unionTypesTest.doubleOrStringAttribute = "foo";
    shouldBeEqualToString('unionTypesTest.doubleOrStringAttribute', 'foo');
    unionTypesTest.doubleOrStringAttribute = undefined;
    shouldBeEqualToString('unionTypesTest.doubleOrStringAttribute', 'undefined');
    unionTypesTest.doubleOrStringAttribute = null;
    shouldBeEqualToString('unionTypesTest.doubleOrStringAttribute', 'null');
    unionTypesTest.doubleOrStringAttribute = {};
    shouldBeEqualToString('unionTypesTest.doubleOrStringAttribute', '[object Object]');
    unionTypesTest.doubleOrStringAttribute = [];
    shouldBeEqualToString('unionTypesTest.doubleOrStringAttribute', '');
    debug('');

    debug('Tests for method arguments');
    shouldBeEqualToString('unionTypesTest.doubleOrStringArg(3.14)', 'double is passed: 3.14');
    shouldBeEqualToString('unionTypesTest.doubleOrStringArg("foo")', 'string is passed: foo');
    shouldBeEqualToString('unionTypesTest.doubleOrStringArg(undefined)', 'string is passed: undefined');
    shouldBeEqualToString('unionTypesTest.doubleOrStringArg(null)', 'string is passed: null');
    shouldBeEqualToString('unionTypesTest.doubleOrStringArg({})', 'string is passed: [object Object]');
    shouldBeEqualToString('unionTypesTest.doubleOrStringArg([])', 'string is passed: ');
    // ToString(Date) can contain timezone, so we can't specify the expectation.
    // Just check that doubleOrStringArg() doesn't throw an exception.
    shouldBeEqualToString('typeof unionTypesTest.doubleOrStringArg(new Date)', 'string');
    shouldThrow('unionTypesTest.doubleOrStringArg()');
    debug('');

    shouldBeEqualToString('unionTypesTest.doubleOrStringArrayArg([])', '');
    shouldBeEqualToString('unionTypesTest.doubleOrStringArrayArg([3.14, "foo"])', 'double: 3.14, string: foo');
    shouldBeEqualToString('unionTypesTest.doubleOrStringArrayArg([1, "foo", "bar", 2])', 'double: 1, string: foo, string: bar, double: 2');
    shouldBeEqualToString('unionTypesTest.doubleOrStringArrayArg([null, undefined, {}, []])', 'string: null, string: undefined, string: [object Object], string: ');
    shouldThrow('unionTypesTest.doubleOrStringArrayArg(null)');
    shouldThrow('unionTypesTest.doubleOrStringArrayArg(undefined)');
    shouldThrow('unionTypesTest.doubleOrStringArrayArg({})');
    debug('');

    shouldBeEqualToString('unionTypesTest.doubleOrStringSequenceArg([])', '');
    shouldBeEqualToString('unionTypesTest.doubleOrStringSequenceArg([3.14, "foo"])', 'double: 3.14, string: foo');
    shouldBeEqualToString('unionTypesTest.doubleOrStringSequenceArg([1, "foo", "bar", 2])', 'double: 1, string: foo, string: bar, double: 2');
    shouldBeEqualToString('unionTypesTest.doubleOrStringSequenceArg([null, undefined, {}, []])', 'string: null, string: undefined, string: [object Object], string: ');
    shouldThrow('unionTypesTest.doubleOrStringSequenceArg(null)');
    shouldThrow('unionTypesTest.doubleOrStringSequenceArg(undefined)');
    shouldThrow('unionTypesTest.doubleOrStringSequenceArg({})');
    debug('');

    debug('Tests for method arguments with defaults');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultDoubleArg()', 'double is passed: 3.14');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultDoubleArg(undefined)', 'double is passed: 3.14');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultDoubleArg("foo")', 'string is passed: foo');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultStringArg()', 'string is passed: foo');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultStringArg(undefined)', 'string is passed: foo');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultStringArg(3.14)', 'double is passed: 3.14');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultNullArg()', 'null is passed');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultNullArg(undefined)', 'null is passed');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultNullArg(null)', 'null is passed');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultNullArg(3.14)', 'double is passed: 3.14');
    shouldBeEqualToString('unionTypesTest.doubleOrStringDefaultNullArg("foo")', 'string is passed: foo');

}
</script>
